Dedupe SpCooldown and IsSitting across components
Remove duplicate SpCooldown from PlayerStateComponent (kept on SpComponent,
widened from int to short) and duplicate IsSitting from PlayerFlagsComponent
(kept on VisualComponent). The bundle's manual pass-through properties are
no longer needed — the source generator now emits natural SpCooldown and
IsSitting accessors since each name is unique.

+++ src/NosCore.GameObject/Ecs/Components/SpComponent.cs
@@ -1,3 +1,3 @@
 namespace NosCore.GameObject.Ecs.Components;
 
-public record struct SpComponent(int SpCooldown, int SpPoint, int SpAdditionPoint);
+public record struct SpComponent(short SpCooldown, int SpPoint, int SpAdditionPoint);
